+++ src/JestProcessManagement/JestProcess.ts
@@ -4,18 +4,18 @@ import { WatchMode } from '../Jest'
 
 export class JestProcess {
   static readonly keepAliveLimit = 5
+  static readonly stopHangTimeout = 500
   private runner: Runner
   private projectWorkspace: ProjectWorkspace
   private onExitCallback: Function
   private jestSupportEvents: Map<string, (...args: any[]) => void>
-  private resolve: Function
+  private stopResolveCallback: Function | null
   private keepAliveCounter: number
   public keepAlive: boolean
-  public stopRequested: boolean
   watchMode: WatchMode
 
   private startRunner() {
-    this.stopRequested = false
+    this.stopResolveCallback = null
     let exited = false
 
     const options = {
@@ -34,10 +34,13 @@ export class JestProcess {
         if (--this.keepAliveCounter > 0) {
           this.runner.removeAllListeners()
           this.startRunner()
-        } else if (this.onExitCallback) {
-          this.onExitCallback(this)
-          if (this.stopRequested) {
-            this.resolve()
+        } else {
+          if (this.onExitCallback) {
+            this.onExitCallback(this)
+          }
+          if (this.stopResolveCallback) {
+            this.stopResolveCallback()
+            this.stopResolveCallback = null
           }
         }
       }
@@ -72,23 +75,31 @@ export class JestProcess {
     this.onExitCallback = callback
   }
 
-  public onJestEditorSupportEvent(event, callback) {
+  public onJestEditorSupportEvent(event: string, callback: (...args: any[]) => void) {
     this.jestSupportEvents.set(event, callback)
     this.runner.on(event, callback)
     return this
   }
 
-  public stop() {
-    this.stopRequested = true
-    this.keepAliveCounter = 1
-    this.jestSupportEvents.clear()
-    this.runner.closeProcess()
+  public stop(): Promise<void> {
     return new Promise(resolve => {
-      this.resolve = resolve
+      this.keepAliveCounter = 1
+      this.stopResolveCallback = resolve
+      this.jestSupportEvents.clear()
+      this.runner.closeProcess()
+
+      // As a safety fallback to prevent the stop from hanging, resolve after a timeout
+      // this is safe since subsequent resolve calls are no-op
+      // TODO: If `closeProcess` can be guarenteed to always resolve, remove this
+      setTimeout(resolve, JestProcess.stopHangTimeout)
     })
   }
 
-  public runJestWithUpdateForSnapshots(callback) {
+  public runJestWithUpdateForSnapshots(callback: () => void) {
     this.runner.runJestWithUpdateForSnapshots(callback)
+  }
+
+  public stopRequested(): boolean {
+    return this.stopResolveCallback !== null
   }
 }