+++ src/extension.ts
@@ -27,6 +27,7 @@ export function activate(context: vscode.ExtensionContext) {
       extension.startProcess()
     }),
     extensionManager.registerCommand(`${extensionName}.stop`, extension => extension.stopProcess()),
+    extensionManager.registerCommand(`${extensionName}.restart`, extension => extension.restartProcess()),
     // dublicate of "show-output" maybe remove?
     extensionManager.registerCommand(`${extensionName}.show-channel`, extension => extension.channel.show()),
     extensionManager.registerCommand(`${extensionName}.coverage.toggle`, extension =>