+++ tests/JestProcessManagement/JestProcessManager.test.ts
@@ -255,11 +255,14 @@ describe('JestProcessManager', () => {
       jestProcessManager.startJestProcess()
       jestProcessManager.startJestProcess()
 
-      jestProcessManager.stopAll()
+      const stopAll = jestProcessManager.stopAll()
 
-      expect(jestProcessMock.mock.instances.length).toBe(2)
-      expect(mockImplementation.stop).toHaveBeenCalledTimes(2)
-      expect(jestProcessManager.numberOfProcesses).toBe(0)
+      expect(stopAll).toBeInstanceOf(Promise)
+      return stopAll.then(() => {
+        expect(jestProcessMock.mock.instances.length).toBe(2)
+        expect(mockImplementation.stop).toHaveBeenCalledTimes(2)
+        expect(jestProcessManager.numberOfProcesses).toBe(0)
+      })
     })
 
     it('does not stop any jest process if none is running', () => {
@@ -270,11 +273,14 @@ describe('JestProcessManager', () => {
 
       jestProcessMock.mockImplementation(() => mockImplementation)
 
-      jestProcessManager.stopAll()
+      const stopAll = jestProcessManager.stopAll()
 
-      expect(jestProcessMock.mock.instances.length).toBe(0)
-      expect(mockImplementation.stop).not.toHaveBeenCalled()
-      expect(jestProcessManager.numberOfProcesses).toBe(0)
+      expect(stopAll).toBeInstanceOf(Promise)
+      return stopAll.then(() => {
+        expect(jestProcessMock.mock.instances.length).toBe(0)
+        expect(mockImplementation.stop).not.toHaveBeenCalled()
+        expect(jestProcessManager.numberOfProcesses).toBe(0)
+      })
     })
   })